fixed some minor things and added component SocialLink as advised in PR's comment
