+++ src/components/UI/SocialLink/index.js
@@ -1,0 +1,3 @@
+import SocialLink from './SocialLink.jsx';
+
+export default SocialLink;
